Fix the non-q8.8 conversions

+++ HTModels/HTModels.h
@@ -10,7 +10,7 @@
 
 //#include <Ht.h>
 
-void relu_forward(int64_t *input_q88_data, int64_t *output_q88_data, size_t numbers );
+void relu_forward(int32_t *input_q88_data, int32_t *output_q88_data, size_t numbers );
